+++ api/notes.json
@@ -1,0 +1,3 @@
+{
+    "notes": []
+}
