+++ index.html
@@ -14,12 +14,14 @@
     <h2 class = "criminalHeading">Glassdale Mystery Page</h2>
     <div class="filters">
         <div class="filter filters__crime"></div>
-    </div>
-    <article class="criminalsContainer"></article>
-    <article class="officersContainer"></article>
-    <div class="filters">
         <div class="filter filters__officers"></div>
     </div>
+    <h2 class="officerHeading">Selected Criminals</h2>
+    <article class="criminalsContainer"></article>
+    <h2 class="criminalHeading">Officers</h2>
+    <article class="noteFormContainer"></article>
+    <article class="criminalsContainer"></article>
+    <article class="officersContainer"></article>
     </header>
     </main>
