+++ styles/criminals.css
@@ -1,5 +1,6 @@
 .criminalHeading{
     text-align: center;
+    margin: 5rem;
 }
 
 .criminalsContainer {
